+++ supervised_benchmarks/tests/test_uci_income.py
@@ -1,13 +1,37 @@
 from pathlib import Path
+from typing import NamedTuple, FrozenSet, Mapping, Literal
 
-from supervised_benchmarks.benchmark import BenchmarkConfig
-from supervised_benchmarks.dataset_protocols import FixedTrain, FixedTest
-from supervised_benchmarks.metrics import get_pair_metric
-from supervised_benchmarks.sampler import FixedEpochSamplerConfig, FullBatchSampler, FullBatchSamplerConfig
-from supervised_benchmarks.uci_income.consts import AnyNetDiscrete, AnyNetDiscreteOut, n_tokens
+from numpy.typing import NDArray
+
+from supervised_benchmarks.dataset_protocols import FixedTrain, FixedTest, DataUnit
+from supervised_benchmarks.ports import Port
+from supervised_benchmarks.protocols import Performer
+from supervised_benchmarks.sampler import FixedEpochSamplerConfig, FullBatchSamplerConfig
+from supervised_benchmarks.uci_income.consts import AnyNetDiscrete, AnyNetDiscreteOut
 from supervised_benchmarks.uci_income.uci_income import UciIncomeDataConfig, uci_income_in_anynet_discrete, \
     uci_income_out_anynet_discrete
-from variable_protocols.variables import var_scalar, ordinal
+from variable_protocols.protocols import Variable
+
+
+class XGBoostModelConfig(NamedTuple):
+    repertoire: FrozenSet[Port]
+
+    ports: Mapping[Port, Variable]
+
+    type: Literal['ModelConfig'] = 'ModelConfig'
+
+    def prepare(self) -> Performer: ...
+
+
+class XGBoostPerformer(NamedTuple):
+    def model(self) -> XGBoostModelConfig:
+        pass
+
+    def perform(self, data_src: DataUnit, tgt: Port) -> NDArray: ...
+
+    def perform_batch(self,
+                      data_src: DataUnit,
+                      tgt: FrozenSet[Port]) -> DataUnit: ...
 
 
 def test_get_data():